Rename modules directory to lib as I feel that better represents what's contained there
