+++ app/lib/palette_extractor.py
